changed as coisas q a gente tinha falado
coloquei os sets pra funcionar a atualizacao das pessoas, mudei as fontes e fiz o esquema das paginas

+++ src/Pessoa.java
@@ -58,7 +58,7 @@ public class Pessoa {
 		return;
 	}
 	
-	public void setFreq(float _freq) {
+	public void setFreq(double _freq) {
 		freq = _freq;
 		return;
 	}
